:recycle: refactor: 상품 삭제 시 자신의 상품이 아닌경우 예외 처리 (#190)

+++ src/main/java/umc/unimade/domain/products/service/ProductsCommandService.java
@@ -45,7 +45,6 @@ public class ProductsCommandService {
     private final ProductRegisterRepository productRegisterRepository;
     private final CategoryRepository categoryRepository;
     private final S3Provider s3Provider;
-    private final SellerRepository sellerRepository;
 
     @Transactional
     public ApiResponse<Void> toggleFavoriteProduct(Long productId, Buyer buyer) {
@@ -129,10 +128,14 @@ public class ProductsCommandService {
     }
 
     // 상품 삭제
-    public void deleteProduct(Long productId) {
+    public void deleteProduct(Seller seller, Long productId) {
 
         Products product = productRepository.findById(productId)
                 .orElseThrow(() -> new ProductExceptionHandler(ErrorCode.PRODUCT_NOT_FOUND));
+
+        if (!product.getSeller().getId().equals(seller.getId())) {
+            throw new ProductExceptionHandler(ErrorCode.PRODUCT_IS_NOT_YOURS);
+        }
 
         productRepository.deleteById(productId);
     }
